+++ jib-core/src/main/java/com/google/cloud/tools/jib/image/json/ImageToJsonTranslator.java
@@ -160,6 +160,8 @@ public class ImageToJsonTranslator {
     }
 
     template.setCreated(image.getCreated() == null ? null : image.getCreated().toString());
+    template.setArchitecture(image.getArchitecture());
+    template.setOs(image.getOs());
     template.setContainerEnvironment(environmentMapToList(image.getEnvironment()));
     template.setContainerEntrypoint(image.getEntrypoint());
     template.setContainerCmd(image.getProgramArguments());